requirements.txt
Flask==3.0.3
mysql-connector-python==9.0.0
nltk==3.8.1
scikit-learn==1.3.2
numpy==1.26.4
scipy==1.11.4
python-dotenv==1.0.1
